Norminette: split quaternion
