+++ glmath/srcs/quaternion.c
@@ -6,7 +6,7 @@
 /*   By: dthalman <daniel@thalmann.li>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/27 15:43:40 by dthalman          #+#    #+#             */
-/*   Updated: 2022/06/01 20:58:03 by dthalman         ###   ########.fr       */
+/*   Updated: 2022/06/09 11:02:39 by trossel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,24 +51,6 @@ void	qion_copy(t_qion *to, const t_qion *copy)
 }
 
 /**
- * @brief effectue le produit de deux quaternion
- * 
- * @param q1 
- * @param q2 
- * @return t_qion 
- */
-t_qion	qion_product(const t_qion *q1, const t_qion *q2)
-{
-	t_qion	q;
-
-	q.w = q1->w * q2->w - q1->x * q2->x - q1->y * q2->y - q1->z * q2->z;
-	q.x = q1->w * q2->x + q1->x * q2->w + q1->y * q2->z - q1->z * q2->y;
-	q.y = q1->w * q2->y - q1->x * q2->z + q1->y * q2->w + q1->z * q2->x;
-	q.z = q1->w * q2->z + q1->x * q2->y - q1->y * q2->x + q1->z * q2->w;
-	return (q);
-}
-
-/**
  * @brief Crée un quaternion de rotation
  * 
  * @return t_qion 
@@ -91,67 +73,6 @@ t_qion	qion_euler_rotation(float x, float y, float z)
 	q.z = c[1] * s[2] * c[3] - s[1] * c[2] * s[3];
 	return (q);
 }
-
-/**
- * @brief effectue une rotation a l'aide des quaternion
- * 
- * @param a 
- * @param rot 
- * @return t_qion 
- */
-t_qion	qion_rotation(const t_qion *q, const t_qion *rot)
-{
-	t_qion	ret;
-	t_qion	inv;
-
-	inv = qion_conjugate(rot);
-	ret = qion_product(rot, q);
-	ret = qion_product(&ret, &inv);
-	return (ret);
-}
-
-t_qion	qion_conjugate(const t_qion *q)
-{
-	t_qion	inv;
-
-	inv = qion_create(q);
-	inv.x *= -1;
-	inv.y *= -1;
-	inv.z *= -1;
-	return (inv);
-}
-
-/**
- * @brief retourne la longueur du quaternion
- * 
- * @param q 
- * @return float 
- */
-float	qion_norm(const t_qion *q)
-{
-	float	norm;
-
-	norm = q->w * q->w + q->x * q->x + q->y * q->y + q->z * q->z;
-	return (sqrt(norm));
-}
-
-/**
- * @brief Retourne un vecteur à partir du point (0,0,0) d'une valeur de 1
- *  exemple, avec le vecteur (2,0,0) on obtiendra (1,0,0)
- *
- * @param v
- */
-void	qion_normalize(t_qion *q)
-{
-	float	n;
-
-	n = qion_norm(q);
-	q->w /= n;
-	q->x /= n;
-	q->y /= n;
-	q->z /= n;
-}
-
 
 t_v3f	qion_rotation_angle(const float angle, const t_v3f *v)
 {
